Add ARecipeProcessor to help abstract recipe processors
